feat: ImageFile Type @types로 이동
