+++ src/screens/home/types.ts
@@ -42,11 +42,3 @@ export interface ModalOption {
   color?: string;
   icon?: ReactNode;
 }
-
-export interface ImageFile {
-  uri: string;
-  type: string;
-  fileName?: string;
-  width?: number;
-  height?: number;
-}
