+++ src/types/ImageFile.ts
@@ -1,0 +1,7 @@
+export interface ImageFile {
+    uri: string;
+    type: string;
+    fileName?: string;
+    width?: number;
+    height?: number;
+}
