+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * as Auth from './AuthDTO';
 export * as Comment from './CommentDTO';
 export * as Feed from './FeedDTO';
 export * as Onboarding from './OnboardingDTO';
+export * from './ImageFile';
